Removing code that won't ever return false
I'm pretty confident this line will never be hit. The reason I think that is that the variable '' is defined using a combination of a call to 'Get-NodeVersions' and 'Get-NodeInstallLocation'. So those are both reading the file system already. The only way I can see this happening is if the version was removed from disk by another process while the cmdlet was executing. So, basically impossible to hit that!